Day 23: Fix bug in recursion
The code still produces the correct solution despite this bug due to the nice structure of the AoC graph. However, for correctness's sake, the bug must be fixed.

This significantly reduces the runtime of the solution. Perhaps we could speed it up again by sharing a singleton cache across all invocations of `maximal_clique`